+++ service/controller/resource/tccp/create_test.go
@@ -197,7 +197,9 @@ func Test_Controller_Resource_TCCP_Template_Render(t *testing.T) {
 				}
 			}
 
-			params, err := r.newParamsMain(tc.ctx, tc.cr, time.Time{})
+			cl := unittest.DefaultCAPIClusterWithLabels(tc.cr.Name, map[string]string{})
+
+			params, err := r.newParamsMain(tc.ctx, cl, tc.cr, time.Time{})
 			if err != nil {
 				t.Fatal(err)
 			}